+++ plugins/Live/vue/src/index.ts
@@ -5,8 +5,8 @@
  * @license https://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
  */
 
-export { default as LiveWidgetRefresh } from './LiveWidget/LiveWidgetRefresh';
+export { default as LiveWidget } from './LiveWidget/LiveWidget.vue';
+export { default as AutoRefreshWidget } from './AutoRefreshWidget/AutoRefreshWidget.vue';
 export { default as TotalVisitors } from './TotalVisitors/TotalVisitors.vue';
 export { default as LivePage } from './LivePage/LivePage.vue';
 export { default as IndexHeader } from './IndexHeader/IndexHeader.vue';
-export { default as LastVisits } from './LastVisits/LastVisits';