+++ tests/PHPUnit/proxy/includes.php
@@ -1,8 +1,7 @@
 <?php
 
 if (!defined('PIWIK_INCLUDE_PATH')) {
-    // NOTE: PIWIK_INCLUDE_PATH must end in '/' or some parts of matomo will break
-    define('PIWIK_INCLUDE_PATH', realpath(dirname(__FILE__) . '/../../..') . '/');
+    define('PIWIK_INCLUDE_PATH', realpath(dirname(__FILE__) . '/../../..'));
 }
 
 if (!defined('PIWIK_USER_PATH')) {
@@ -14,7 +13,7 @@ if (!defined('PIWIK_TEST_MODE')) {
 
 $GLOBALS['MATOMO_PLUGIN_DIRS'] = array(
     array(
-        'pluginsPathAbsolute' => realpath(PIWIK_INCLUDE_PATH . 'tests/resources/custompluginsdir'),
+        'pluginsPathAbsolute' => realpath(PIWIK_INCLUDE_PATH . '/tests/resources/custompluginsdir'),
         'webrootDirRelativeToMatomo' => '../../resources/custompluginsdir'
     ),
 );